+++ tests/tf1/functional/test_metric_ppl_genmodel.py
@@ -33,7 +33,7 @@ class TestMetricPplGenModel(unittest.TestCase):
         res_fidelity = json_decode_string(res_fidelity.stdout.decode())
         print('Fidelity PPL result:', res_fidelity, file=sys.stderr)
 
-        self.assertAlmostEqual(res_fidelity[KEY_METRIC_PPL_MEAN], 2560.187255859375, delta=1e-5)
+        self.assertAlmostEqual(res_fidelity[KEY_METRIC_PPL_MEAN], 2560.187255859375, delta=2e-1)
 
 
 if __name__ == '__main__':